+++ SokobanLogique/Movable.h
@@ -7,7 +7,7 @@ namespace Sokoban
 		public Case
 	{
 	public:
-		Movable(size_t x, size_t y, size_t z) : Case(x,y,z) {}
+		Movable(int x, int y, int z) : Case(x,y,z) {}
 		virtual bool canMove(Direction) const=0;
 		virtual Direction applyMove();
 		virtual bool move(Direction)=0;
